To convert from infusion/intravenous administration to first-order oral absorption

To convert from infusion/intravenous administration to first-order oral absorption

Usage

addDepot(
  model,
  central = "central",
  depot = "depot",
  absRate = "ka",
  lag = paste0("lag", depot),
  lagIni = NA,
  fdepotIni = NA,
  absRateIni = 1
)

Arguments

model

The model as a function (or something convertible to an rxUi object)

central

central compartment name

depot

depot compartment name

absRate

absorption rate parameter name

lag

lag parameter name

lagIni

Initial value for the lag time (`NA` to omit)

fdepotIni

Initial value for the depot bioavailability (`NA` to omit)

absRateIni

Initial value for the first order rate

Value

a model with the depot added

Examples

# most of the examples in the model library already have a depot.
# for this example we will remove the depot and then add it back
readModelDb("PK_1cmt_des") |>
  removeDepot() |>
  addDepot()
#>  
#>  
#> ! remove population parameter `lka`
#> ℹ promote `lka` to population parameter with initial estimate 1
#> ℹ change initial estimate of `lka` to `1`
#>  ── rxode2-based free-form 2-cmt ODE model ────────────────────────────────────── 
#>  ── Initalization: ──  
#> Fixed Effects ($theta): 
#>    lka    lcl    lvc propSd 
#>   1.00   1.00   3.45   0.50 
#> 
#> States ($state or $stateDf): 
#>   Compartment Number Compartment Name
#> 1                  1            depot
#> 2                  2          central
#>  ── Model (Normalized Syntax): ── 
#> function() {
#>     description <- "One compartment PK model with linear clearance using differential equations"
#>     dosing <- c("central", "depot")
#>     ini({
#>         lka <- 1
#>         label("First order absorption rate (ka)")
#>         lcl <- 1
#>         label("Clearance (CL)")
#>         lvc <- 3.45
#>         label("Central volume of distribution (V)")
#>         propSd <- c(0, 0.5)
#>         label("Proportional residual error (fraction)")
#>     })
#>     model({
#>         ka <- exp(lka)
#>         cl <- exp(lcl)
#>         vc <- exp(lvc)
#>         kel <- cl/vc
#>         d/dt(depot) <- -ka * depot
#>         d/dt(central) <- ka * depot + -kel * central
#>         Cc <- central/vc
#>         Cc ~ prop(propSd)
#>     })
#> }