Add residual error to a model

Add residual error to a model

Usage

addResErr(model, reserr)

Arguments

model

The model as a function

reserr

The type or types of residual error (currently "addSd", "propSd", and "lnormSd" are accepted)

Value

The model with residual error modified

Details

For reserr, the parameter will be named with the dependent variable from the model as a prefix. For example, if the dependent variable in the model is cp, the parameter name for propSd will become cppropSd.

Examples

readModelDb("PK_1cmt") %>% addResErr("addSd")
#>  
#>  
#> ! remove population parameter `propSd`
#> ℹ add residual parameter `cpaddSd` and set estimate to 1
#> ℹ change initial estimate of `cpaddSd` to `1`
#> function () 
#> {
#>     description <- "One compartment PK model with linear clearance"
#>     ini({
#>         lka <- 0.45
#>         label("Absorption rate (Ka)")
#>         lcl <- 1
#>         label("Clearance (CL)")
#>         lvc <- 3.45
#>         label("Central volume of distribution (V)")
#>         cpaddSd <- c(0, 1)
#>     })
#>     model({
#>         ka <- exp(lka)
#>         cl <- exp(lcl)
#>         vc <- exp(lvc)
#>         cp <- linCmt()
#>         cp ~ add(cpaddSd)
#>     })
#> }
#> <environment: 0x55d3048bc8e8>
readModelDb("PK_1cmt") %>% addResErr("lnormSd")
#>  
#>  
#> ! remove population parameter `propSd`
#> ℹ add residual parameter `cplnormSd` and set estimate to 1
#> ℹ change initial estimate of `cplnormSd` to `0.5`
#> function () 
#> {
#>     description <- "One compartment PK model with linear clearance"
#>     ini({
#>         lka <- 0.45
#>         label("Absorption rate (Ka)")
#>         lcl <- 1
#>         label("Clearance (CL)")
#>         lvc <- 3.45
#>         label("Central volume of distribution (V)")
#>         cplnormSd <- c(0, 0.5)
#>     })
#>     model({
#>         ka <- exp(lka)
#>         cl <- exp(lcl)
#>         vc <- exp(lvc)
#>         cp <- linCmt()
#>         cp ~ lnorm(cplnormSd)
#>     })
#> }
#> <environment: 0x55d306cd33b0>
readModelDb("PK_1cmt") %>% addResErr(c("addSd", "propSd"))
#>  
#>  
#> ! remove population parameter `propSd`
#> ℹ add residual parameter `cpaddSd` and set estimate to 1
#> ℹ add residual parameter `cppropSd` and set estimate to 1
#> ℹ change initial estimate of `cpaddSd` to `1`
#> ℹ change initial estimate of `cppropSd` to `0.5`
#> function () 
#> {
#>     description <- "One compartment PK model with linear clearance"
#>     ini({
#>         lka <- 0.45
#>         label("Absorption rate (Ka)")
#>         lcl <- 1
#>         label("Clearance (CL)")
#>         lvc <- 3.45
#>         label("Central volume of distribution (V)")
#>         cpaddSd <- c(0, 1)
#>         cppropSd <- c(0, 0.5)
#>     })
#>     model({
#>         ka <- exp(lka)
#>         cl <- exp(lcl)
#>         vc <- exp(lvc)
#>         cp <- linCmt()
#>         cp ~ add(cpaddSd) + prop(cppropSd)
#>     })
#> }
#> <environment: 0x55d30870ccb8>